Move the implementation of the FlowLogicRefFactory internal to the node as it is an implementation detail, not an API.
