+++ node/src/test/java/net/corda/node/services/events/FlowLogicRefFromJavaTest.java
@@ -1,8 +1,14 @@
-package net.corda.core.flows;
+package net.corda.node.services.events;
 
-import org.junit.*;
+import net.corda.core.flows.FlowLogic;
+import net.corda.core.flows.FlowLogicRefFactory;
+import net.corda.node.services.statemachine.FlowLogicRefFactoryImpl;
+import org.junit.Test;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
 
 public class FlowLogicRefFromJavaTest {
 
@@ -55,7 +61,7 @@ public class FlowLogicRefFromJavaTest {
         argsList.add(ParamType1.class.getName());
         argsList.add(ParamType2.class.getName());
         whiteList.put(JavaFlowLogic.class.getName(), argsList);
-        FlowLogicRefFactory factory = new FlowLogicRefFactory(whiteList);
+        FlowLogicRefFactory factory = new FlowLogicRefFactoryImpl(whiteList);
         factory.create(JavaFlowLogic.class, new ParamType1(1), new ParamType2("Hello Jack"));
     }
 
@@ -63,7 +69,7 @@ public class FlowLogicRefFromJavaTest {
     public void testNoArg() {
         Map<String, Set<String>> whiteList = new HashMap<>();
         whiteList.put(JavaNoArgFlowLogic.class.getName(), new HashSet<>());
-        FlowLogicRefFactory factory = new FlowLogicRefFactory(whiteList);
+        FlowLogicRefFactory factory = new FlowLogicRefFactoryImpl(whiteList);
         factory.create(JavaNoArgFlowLogic.class);
     }
 }
