+++ node/src/test/kotlin/net/corda/node/services/MockServiceHubInternal.kt
@@ -4,7 +4,6 @@ import com.codahale.metrics.MetricRegistry
 import net.corda.core.crypto.Party
 import net.corda.core.flows.FlowInitiator
 import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.FlowLogicRefFactory
 import net.corda.core.node.NodeInfo
 import net.corda.core.node.services.*
 import net.corda.core.transactions.SignedTransaction
@@ -13,6 +12,7 @@ import net.corda.node.serialization.NodeClock
 import net.corda.node.services.api.*
 import net.corda.node.services.messaging.MessagingService
 import net.corda.node.services.schema.NodeSchemaService
+import net.corda.node.services.statemachine.FlowLogicRefFactoryImpl
 import net.corda.node.services.statemachine.FlowStateMachineImpl
 import net.corda.node.services.statemachine.StateMachineManager
 import net.corda.node.services.transactions.InMemoryTransactionVerifierService
@@ -30,7 +30,7 @@ open class MockServiceHubInternal(
         val mapCache: NetworkMapCacheInternal? = MockNetworkMapCache(),
         val scheduler: SchedulerService? = null,
         val overrideClock: Clock? = NodeClock(),
-        val flowFactory: FlowLogicRefFactory? = FlowLogicRefFactory(),
+        val flowFactory: FlowLogicRefFactoryInternal? = FlowLogicRefFactoryImpl(),
         val schemas: SchemaService? = NodeSchemaService(),
         val customTransactionVerifierService: TransactionVerifierService? = InMemoryTransactionVerifierService(2)
 ) : ServiceHubInternal() {
@@ -56,7 +56,7 @@ open class MockServiceHubInternal(
         get() = throw UnsupportedOperationException()
 
     override val monitoringService: MonitoringService = MonitoringService(MetricRegistry())
-    override val flowLogicRefFactory: FlowLogicRefFactory
+    override val flowLogicRefFactory: FlowLogicRefFactoryInternal
         get() = flowFactory ?: throw UnsupportedOperationException()
     override val schemaService: SchemaService
         get() = schemas ?: throw UnsupportedOperationException()
